+++ src/containers/MyPage/index.js
@@ -3,4 +3,6 @@ export { default as MyComment } from "./MyComment";
 export { default as MyFeed } from "./MyFeed";
 export { default as MyInfo } from "./MyInfo";
 export { default as MyPageIntro } from "./MyPageIntro";
+export { default as MyPages } from "./MyPages";
 export { default as MyStudy } from "./MyStudy";
+export { default as ProfileImage } from "./ProfileImage";
